\name{xhare}
\alias{xhare}
\title{Hare: hazard regression}
\description{Driver function for \code{\link{dhare}}, \code{\link{hhare}}, \code{\link{phare}}, \code{\link{qhare}},
and \code{\link{rhare}}. This function is not intended for use by itself.}
\usage{xhare(arg1, arg2, arg3, arg4) }
\arguments{
\item{arg1,arg2,arg3,arg4}{ arguments.  }
}
\details{ This function is used internally.  }
\note{This function is not intended for direct use.}
\references{Charles Kooperberg, Charles J. Stone and Young K. Truong (1995).
Hazard regression.  \emph{Journal of the American Statistical
Association}, \bold{90}, 78-94.

Charles J. Stone, Mark Hansen, Charles Kooperberg, and Young K. Truong.
The use of polynomial splines and their tensor products in extended
linear modeling (with discussion) (1997).  \emph{Annals of Statistics},
\bold{25}, 1371--1470.}
\author{ Charles Kooperberg \email{clk@fhcrc.org}.}
\seealso{ \code{\link{hare}},
\code{\link{dhare}}, \code{\link{hhare}}, \code{\link{phare}}, \code{\link{qhare}},
\code{\link{rhare}}.}
\keyword{distribution}
\keyword{smooth}
\keyword{survival}
